unpack.h: implemented unpack9x32
This is an attempt to implement optimal routines unpackMxN for decoding
m-bit numbers, n of them, base64-encoded.  The encoded block must start
on a character boundary, and should preferably span the whole number of
characters (i.e. a multiple of 6 bits), as is the case with unpack9x32.
But some routines may provide a few extra bits for an alternative /
variable-length bitstream.

This is primarily based on the two papers co-authored by Daniel Lemire:
- Faster Base64 Encoding and Decoding Using AVX2 Instructions
  https://arxiv.org/abs/1704.00605
- Decoding billions of integers per second through vectorization
  https://arxiv.org/abs/1209.2137

The code is untested.