factored base64glue12 and base64glue24
After 6-bit values are decoded, they are transformed into 6+6=12 and
then 12+12=24 lanes with two multiplications.  Multiplication latency
is 5 cycles, and the second multiplication cannot start before the first
completes.  But note that we do not always have to wait for full 24-bit
lanes, sometimes we can use fewer bits.  In 24-bit lanes, the bits are
spread as follows:

    9+9+6 | 3+9+9+3 | 6+9+9

For the first load, we can provide 9 bits after 6+6=12 multiplication.
For the second and third loads, we can provide 3 resp. 6 bits
immediately.  Useful things can be done in parallel with multiplication.

unpack9x32: 0.91 cycles per integer