sketched unpack30x3
Let's try the other end - 30-bit numbers.  This means the average delta
is about 2^30, so there are very few of those numbers, most probably 1..4.
To decode 3 numbers, we need to load 30*3/6=15 characters, which makes
sense: we can load 16 characters and provide extra 6 bits for the q-bitstream.
Note that there must be that extra character if there is a q-bitstream.
To decode 2 numbers though, we would need to load 10 characters, and how
do you even load 10 characters into xmm.  So 30x3 seems the only viable
option.

The problem is then how to extract those 6 bits.  We use movmskps which
extracts the sign bit from each of the 4 lanes, after rearranging 4 24-bit
lanes into 3 32-bit lanes.  This means that the encoder is gonna be more
complicated: it cannot simply pack 30*3 bits into 15 characters, those
extra 6 bits must be known in advance.