pack.h: factored base64unglue
This routine is somewhat inefficient.  Daniel Lemire describes a better
routine which uses two multiplications to shift the data within 16-bit
lanes.  I believe this technique only applies to the original/standard
base64 bitstream which specifies big-endian byte order.  The downside
is that the data must be reshuffled symmetrically in the decoder.
This boils down to just changing the base64glue12 constant from
0x40014001 to 0x01400140, and likewise in base64glue24.  But this also
means that we cannot use the lower 6 bits before gluing them.