implemented 27x4c18
This demonstrates that sometimes we can successfully combine SIMD with
scalar lookup.  This works best only when the numbers come out right,
but this is exactly the case.  When carefully prepared, the last part
rocks and is nothing short of amazing.

    int lo = base64dec2(s + 16);
    if (lo < 0) return false;
    v[3] |= lo;

No extra ops on the scalar value!  So it runs a bit faster than 27x3.

unpack27x3: 3.53 cycles per integer
unpack27x4: 3.33 cycles per integer

The only problem is that it isn't clear if we'll ever need 27x4.
With big m, it's the smaller the better.