neon: implemented unpack29x3c15e3
NEON doesn't provide an instruction to pack sign bits into a
general-purpose register, so how do we extract three extra bits?
Well, we need to unpack three integers, but we store four, and
the sign bits come from the fourth lane.  The fourth lane has been
glued, so we need to extract bits #5, #11, #17.  This can be done
with multiplication, as explained in
https://stackoverflow.com/questions/14547087/extracting-bits-with-a-single-multiplication/14551792#14551792

      (and
        (= ((_ extract 17 17) x) ((_ extract 31 31) y))
        (= ((_ extract 11 11) x) ((_ extract 30 30) y))
        (= ((_ extract  5  5) x) ((_ extract 29 29) y))