implemented 25x3c13e3o1
Getting the extra bits with NEON is a bit hard: when the extra lane is
filled with bytes (5,9,14,-1), the equation for bits #5,#15,#17 has no
solution.  There is a solution if you move one byte up: (5,9,-1,14), but
then the highest bit that we need is masked out/shifted away.  Instead,
we need to break the order: (14,5,9,-1), and remap the bits.

        (= ((_ extract 23 23) x) ((_ extract 30 30) y))
        (= ((_ extract 13 13) x) ((_ extract 29 29) y))
        (= ((_ extract  1  1) x) ((_ extract 31 31) y))