implemented 23x4c16e4
This time the extra bit is the LSB bit in each lane.  This lets us trade
mask for shift.  The problem with mask is that it has to be loaded from
memory, and it takes a register (if the load is factored out of a loop).
Thus SMID shift by immediate should be preferred to mask.  (There are
some older CPUs, notably Xbox 360 Xenon, on which variable shifts are
slow; shift by immediate should be fast everywhere.)

Note that the extra bits are available at the right position immediately
after unpack6, so to issue movmskps, we don't have to wait for glue12+24.
With NEON, we can no longer put the extra bits in the spare lane, so we
need a separate move (disguised under y[0]).