implemented 22x4c15e2o1
The two extra bits are stored as LSB bits in the first lane, and can be
extracted with movd and mask.  This is the main reason why the primitive
is marked with o1 - we can't discard the first byte and still extract
the two bits with movd and a single op.  (Note that pextrb is more
expensive than movd, and the second op may be required to shift the
bits.)

Or can we?  Perhaps we can discard the first byte and extract the two
MSB bits with a single shift.  We'll see about that.