restrict the pointers in LOOP_##unpack
Annotations like "restrict" and "unlikely" should be added at the
topmost level.  The compiler then should propagate them to the inline
function.  Interestingly enough, restricting the pointers in
LOOP_##unpack does make a difference, but doing so in wrap_LOOP
does not.  It seem that GCC cannot propagate restricted pointers
when both the pointers and the inline function are passed as arguments
to another inline function.  I've always suspected that GCC sucks,
even if only a little bit.  Our darling sucky-sucky compiler.