lift the o1 restriction on 22x4c15e2
We need to discard the first byte, and the two extra bits must be the
MSB bits of the first lane (so that we can grab them with movd and shr).
After the bits in the first lane are glued, we subdivide them as
follows: 24=6|10|6|2, where 10|6 are the bits to form the values and
the vertical bar indicates an 8-bit boundary.  In other words, we can
stick 10| to other bytes on the left and |6 on the right.  This makes
it a contiguous stream which is then processed as usual.

The scalar routine comes out quite nice.  The value in the first lane
is 10+12 bits, i.e. easily completed with 2 characters.  So there is no
complex bit shuffling, the scalar routine is almost as good, but not
perfectly so:

-unpack22x4: 4.88 cycles per integer
+unpack22x4: 5.03 cycles per integer

I tried another optimization: trading shift for movzx (aka movzwl).
Modern CPUs can isolate 16 or 8 bits without an ALU op.  It didn't
work well in this case though, it breaks the dec2/dec3 pattern and
requires another (x < 0) check.