optimize sse4 unpack21x4c14
The purpose is to avoid multiplication, which is somewhat expensive.
The optimization is possible because 21+21 is a multiple of 6, i.e. we
can get 64-bit lanes with identical bit-level layout, which can obviate
the need to shift each 32-bit lane by a different amount.  To get the
identical layout, we reshuffle the bytes before gluing them.  We then
get 66633 33666 bit pattern per 64-bit lane.  We mask out the high 3
bits in the first lane, shift away the 3 low bits in the second lane,
and blend the result.

-unpack21x4: 2.38 cycles per integer
+unpack21x4: 2.24 cycles per integer