implemented 20x4c14e4(b) with movd
This is an alternative 20x4 version: extra bits are concentrated in the
first lane and can be extracted with movd.  The routine discards two
leading bytes, so the contents of the first lane, when glued, are 12|4|8,
where 4 are the extra bits and 8 are the value bits.  So with SSE4, we
need two additional ops to extract the 4 bits.  It is actually possible
to change the layout to 12|8|4 and extract the extra bits with a single
op, which is shown in the alternative code path (under #else).  But this
significantly slows down the scalar routine.  This is because the 8 bits
in the 12|8|4 layout cross an 8-bit boundary, and have to be distributed
among two lanes.

Like the (a) version, this (b) version doesn't need to use different
shifts for each lane, which can only be emulated with multiplication.
This is because 20+20 is a multiple of 8, so we can make 64-bit lanes
which have identical bit-level layout.  We can further adjoin the bytes
to the lane's upper boundary, which lets us trade mask for shift.