combine 12+6 bits in 18x5c15
Since we have full control over gluing, we can glue just a little bit:
6+6=12, and then combine some of the 12 bits with some of the remaining 6.
We then get the lanes with 12+(6+0) bits, to which glue24 is applied.
The only problem is with the first lane, which is offloaded early with
12+12 bits, of which 6 leading bits are to be discarded.  We fix it with
a few ALU ops.

The improvement is modest on Haswell:
-unpack18x5: 2.33 cycles per integer
+unpack18x5: 2.28 cycles per integer

The improvement is more noticeable on Opteron 1100 (Cortex-A57),
although the scalar version is still a bit faster.
-unpack18x5: 0.86 cycles per integer
+unpack18x5: 0.81 cycles per integer

The scalar version is as fast as it was, it just combines the bytes
in a slightly different order.

Update: it seems to be quite a bit slower on Ryzen 2700, although it is
hard to measure the exact time because of its hyper-trolling.  Don't
know what to do. :(
-unpack18x5: 1.52 cycles per integer
+unpack18x5: 1.68 cycles per integer