Revert "combine 12+6 bits in 18x5c15"
Multiplication is the culprit: pmulld latency on Haswell is 10 cycles,
not 5 cycles as I thought.  (Other multiplication instructions, such as
pmaddubsw and pmaddwd used in glue12 resp. glue24, indeed have 5-cycle
latency.)  So the earlier version should be preferred, unless pmulld is
slow.  It seems that Intel's been reluctant to improve pmulld, its
latency in Skylake is still 10.  On the other hand, AVX2 introduces
variable shifts, specifically vpsrlvd, whose latency is 2/1/3 cycles
on Haswell/Skylake/Ryzen.  We may be better off with a separate AVX2
version.