implemented 15x6c15
This reintroduces the technique of blending: when there is only one lane
to be shifted, and variable shift is backed by multiplication, we can
shift by immediate and blend.  This is a significant improvement on
Haswell where pmulld is slow.  It is likely to be a small improvement
even if multiplication is fast.  Trading multiplication for two cheaper
instructions is a no-brainer.

-unpack15x6: 2.33 cycles per integer
+unpack15x6: 2.07 cycles per integer