try to optimize unpack13x6c13o1
This trades VSHL32V (+ 128-bit constant) for 3 cheap instructions.
Generally this shouldn't be profitable: on Ryzen 2700, the SSE4 version
(with 3 cheap instructions) and the AVX2 version (with true vpsllvd) run
at about the same speed.  On Haswell however, multiplication is slow,
and the profit is sizable:

-unpack13x6: 2.36 cycles per integer (pmulld)
+unpack13x6: 2.29 cycles per integer (3 cheap instructions)

This probably means that we need to discern these two trade-offs:
we may want to trade VSHL32V+128C for 2 instructions even on AVX2,
but only SSE4:pmulld is worth 3 instructions.