reimplemented 9x32c48
It's very good, and it shows that vertical packing can be much faster
than the VSHLV32 approach.  But it can only handle long blocks, because
ideally the combined width of each lane must be both a multiple of m
and a multiple of 24 (in this case 24+24+24, a multiple of 9).

unpack9x10: 1.72 cycles per integer
unpack9x32: 0.95 cycles per integer

Perhaps we may have two loops: the 9x32 loop handles big blocks, and
then the 9x10 loop handles just a few smaller blocks.  The loops may
also differ in how they handle the q-bitstream: the bigger loop may read
the bitstream in 24-bit chunks, while the small loop only in 12-bit
chunks.