setstring.h: designed the API
This is in part based on rpmss.h, although I wrote it mostly from
scratch, no copy-pasting took place.  There is a tough decision: the
encoding routines expose the m parameter, although we generally do not
expect the caller to vary m.  But otherwise we would need to calculate m
for the second time, and its calculation is non-trivial.  The decoding
routines do not expose m.

Still, this makes some sense: to obtain a shorter set-string, we
sometimes try two m values, in which case setstring_encinit returns
negative m.  If the caller wants to disable the optimization, he can
pass abs(m) instead of m to setstring_encode.