setstring-enc.c: implemented setstring_encinit
This is in part based on rpmss.c and the more recent setstring64.c,
although I gave it some more thought.  The equivalence

	maxbits(m) >= maxbits(m - 1)  iff  m >= log2(dv)

is quite remarkable.  Unfortunately I now have some more optimizations
on my mind which can invalidate this condition.  One idea is that
sometimes we can strip the last character without losing information.
E.g. if the last character holds only one r-bit, then with the probability
of 1/2 it is '0'.  If the last character holds a few r-bits, but they're
all zero, we can still discard it.  If the last character holds q-bits
(which complement a SIMD block), we can skip a few q-bits if they're all
1 (stop bits).  But this would boosts variation greatly, so it would
make sense to try m < log2(dv) in hopes of dropping the last character,
if it didn't drop with m = log2(dv).

So I'm not sure whether this single enc_maxbits() invocation will enter
the final version.  But I also find it hard, if at all possible, to
think over everything in advance.  Mozart is said to have composed
symphonies in his head, but realistically, you can't build edifices
in your head and then claim it to be a preferred construction method.
Getting a first working version at this stage should be a priority.

Another possible optimization is that, when we need to encode only a few
values, we can implement a special mode with m == bpp.  In other words,
there will be no q-bits.  The resulting string should be preferred if
it is not longer than the alternative strings with m < bpp.