setstring-dec.c: sketch the decoding routine
This is a much more elaborate version of the previous sketch made on
Nov 11 (that didn't make it into the git history).  It was communicated
to me by the Higher Intelligence.  The idea is that the data format is
defined by the decoder.  Specifically the decoder defines the exact
condition when to switch from bulk decoding with SIMD blocks to the tail
loop that reads everything from the bitstream.  The condition must be
tight, but also must be easy to compute: it involves division by 6, the
compiler replaces it with multiplication.  The condition is not the
tightest possible though, because it doesn't take into account the bits
in the bitstream register.

The only missing part is where to plug ko, currently the unused
parameter which indicate that some unpack() routines overread 1 byte.