setstring-enc.c: sketch the encoder
Finally it seems to work!  (I'm able to encode and decode an array of
20-bit numbers with m=10, using unpack10x9c15.  I also verified with
printf debugging that the condition in the loop control is the same
across the encoder and the decoder.)

The encoder is less straightforward than the decoder, because each SIMD
block has two demands:
- In order to write the block, we also need to supply extra ke bits.
  These bits only appear during the processing of the next block.
- Before the block is written, we must write q-segments that would
  complete the values for the previous block.

In the general case, this is best done with a queue of delta blocks.
We need to calculate these delta blocks anyway, in order to pass them
to the packing routine.  Thus the queue thinly disguises the stack space
where we store the deltas.