bench-unmet.c: imitate the cache for Provides
When everything gets decoded, performance mainly depends on long
set-versions that come on behalf of Provides.  In reality Provides
versions are cached, so short versions are somewhat important as well.
This introduces another twist (found in rpmss but not in set.c):
very short Provides version should not be cached, because searching
the cache has its cost.

11.0 cycles per integer