trying to implement two q-bitstreams
This is a study of how two q-bitstream might work, and their impact
on the decoder performance.  The idea is that the performance of the
q-delta loop is bounded by the loop-carried data dependency:

    z = ctz32(b), z++, b >>= z;

The next iteration cannot start (ctz32) before the last instruction
(shr cl) is finished.  On Haswell, the latency of ctz32 is 3 cycles,
and shr cl is 2.  Thus the loop cannot progress faster than 6 cycles
per integer, even if the refill of b is amortized well, and assuming
that the delta stage is executed in parallel / does not defer the next
iteration.  Note that simply unrolling the loop won't help.  Instead,
we need to run two independent chains in parallel:

    z0 = ctz32(b0), z0++, b0 >>= z0;
    z1 = ctz32(b1), z1++, b1 >>= z1;

This of course affects the data format: SIMD blocks are now interleaved
with the segments of two independent q-bitstreams.

The implementation is provisional, and has some shortcomings:
- q-segments are refilled with full 5 characters.  Whenever we need
  need another q value, the characters must be there.  This affects the
  loop control: it is offset by an arbitrary constant of 66 so that
  the tail loop is invoked sooner.  If we want the same loop control
  (in order to process as many SIMD blocks as possible), we need to
  handle the situation when e.g b0 is at its end but b1 is not,
  so the last few q bits must be read from b1.  This is easier done
  in the decoder than in the encoder.  (The decoder just reads the
  bits from where they are left, but the encoder must recognize a few
  special conditions and interleave the bits from two q-bitstreams
  plus the tail bitstream properly.)
- The encoder uses U128 integer type to hold variable-length snippets.
  This of course may (and will) overflow, but the chances of this are
  low enough, so the overflow doesn't happen with our data set.  (The
  size is not limited because long stretches of zeroes are possible.
  The U128 type should be replaced with a fixed-size structure which
  counts the zeroes without materializing them too soon.)

Performance on Haswell:
-9.3 cycles per integer
+8.8 cycles per integer

With -R:
-12.6 cycles per integer
+11.9 cycles per integer

Short versions do not benefit, because there are fewer SIMD blocks.
(This will hold somewhat true even with the tight loop control, simply
because the second bitstream is read in a greedy manner.)  So let's
study the effect on long versions separately:

    $ zstd -d <conv.setcmp.zst |awk 'length($1)>=1024{print$1,$1}' |
      ./bench-unmet-sse4

Haswell:
-8.5 cycles per integer
+7.5 cycles per integer

Ryzen 2700:
-6.2 cycles per integer
+5.1 cycles per integer

Not much, but not negligible either.  Not sure what to do with it.