base64pack: implemented 20x6c20
pack20x4c14e4 cannot be used with two q-bitstreams, because the number
of extra decoded bits (e4) cannot exceed the number of bits that are
guaranteed to be consumed in the q-delta loop (otherwise the uint32_t
register may overflow).  To complement 4 decoded values (x4) from a
single q-bitstream, we need to consume at least 4 stop-bits, so it works
fine.  But with two bitstreams, the extra 4 bits are placed back into
one of the two bitstreams; only 2 bits are then guaranteed to be
consumed from that bitstream, and two from the other.