cache.c: hash16 and linear search
SIMD search is measurably faster than its scalar variant, but only with
16-bit hashes.  I first tried SIMD search on 32-bit hashes (a long time
ago) and decided that it wasn't much faster and wasn't worth the trouble.
Then came the insight that 16 bits are just enough to make a good search
function:

- With 16-bit hash, we expect very few collisions among sqrt(2^16)=256
  hash values, as predicted by the birthday paradox.
- Four base64 characters provide only 6*4=24 bits of entropy, so a full
  32-bit hash should be more complex (perhaps taking two loads).
- If we hash only a few bytes (instead of the whole string), there will
  be collisions anyway.  This is because two versions that differ only
  in a few elements are likely to begin with the same characters.

$ zstd -d <conv.setcmp.zst |awk 'length($1)>=128{print length($1),$1}' |
  sort -u |wc -l
2651
$ zstd -d <conv.setcmp.zst |awk 'length($1)>=128{print length($1),$1}' |
  sort -u |sed 's/set:\(........\).*/\1/p' |sort -u |wc -l
2473

In rpmss/, I also have a NEON variant, but it runs slower on those few
older ARM cores that I tried.

To make the search even faster, we could resort to hardware designs,
e.g. to partition the cache into 8 sets, each set 32 ways.  With 3+16
hash bits, we may first select the right set, and then search 32 hash
values in O(1) time.  But this would be overengineering, because of the
cost structure: the full cost of cache_decode() is dominated by decode10
(80%) and memcmp (11%), then some decode9, malloc+free, and memcpy.
The linear search costs only about 1%.