cache.c: the LRU move primitive
This is new stuff.  I first tried to limit the size of LRU move in rpmss
last year, and now I realized that the update and the insertion can be
unified and handled with the same primitive.

This new LRU move is quite cheap, about 3x cheaper than linear search.
Compared to moving items to the very front, it also improves the hit
ratio by about 0.5%, from 76.47% to 76.97% in one particular setting
that I'm studying.