cache.c: Putin it all together
The length isn't stored in cache_ent.  There's no need to compare
the length, we just invoke memcmp with the caller's length.   Note that
memcmp may overread a few extra bytes from ENT_V before it detects a
mismatch near the end of string.  This is reasonably safe, because v[]
is supposed to be reasonably long (short strings should not be cached).
The only way memcmp() can misbehave and read past the end of an entry
if the caller passes the string with an embedded zero.  But then it can
be argued that the caller has violated the precondition (didn't pass a
valid length), so all bets are off.