setcmploop.c: initial revision
Now that the decoding speed has been improved by about 2x, setcmploop
will be one of the two most expensive routines.  Specifically, valgrind
shows that decode10 takes more instructions, but "perf report" suggests
that setcmploop takes more time to execute.  This is probably because
setcmploop has an inner loop that that finds Pv such that *Pv >= Rval,
so roughly it takes one branch misprediction perl R[] element (assuming
that the branch predictor always predicts the next iteration, which
would arguably be the most sensible prediction).

Of course, you can rewrite the loop without unpredictable branches,
replacing them with cmov instructions.  So I came up with the following
assembly:

    1:
	mov (Rv),Rval
	cmp Rval,(Pv)
	lea 4(Rv),Addr
	cmovae Addr,Rv // increase Rv if *Pv >= *Rv
	lea 4(Pv),Addr
	cmovbe Addr,Pv // increase Pv if *Pv <= *Rv
	cmova Zero,Ge
	cmovb Zero,Le
	cmp Rv,Rend
	je 2f
	cmp Pv,Pend
	jne 1b
    2:
	cmp Pv,Pend
	cmovne Zero,Le
	cmp Rv,Rend
	cmovne Zero,Ge

But it runs much slower, a few times slower than setcmploop.c.  The
thing is, mispredicted branches are not super-expensive, neither are
cmovs super-cheap.  The approach that works best is a refinement
of what I developed a long time ago: to step two elements at a time,
in a 3x unrolled loop (i.e. to probe Pv[2], Pv[4] and Pv[6]), and
to step back with a single cmov.

gcc produces very efficient output for this code.  I tried to write
my own assembly, but I couldn't beat gcc!  clang is not that good,
but not that bad either.  (I only have clang-7.)

It seems that gcc produces good code in part due to luck.  I tired to
add likely() annotations and reorder if-break statements to improve
the while(1) loop - it gets worse.

I also tried to step 4 elements at a time, and bisect backwards with
two cmovs (similar to the way it's implemented in rpmss, and indeed
in set.c).  It didn't work much faster even with Pn/Rn >= 64.