downsample.c: initial revision
Very similar to set.c, with one subtle optimization: v1val can be
preloaded without boundary check (because v1end < v2end).  As a result,
if-break becomes the last statement, and can be used to transfer control
to the beginning of the loop.  This can't be done with v2val (unless we
install a sentinel for all decoded versions), so it takes a separate jmp
instruction to start a new iteration (which is a tad bit slower).

    if (v1val < v2val) {
	*w++ = v1val;
	v1++;
	v1val = *v1;
	if (v1 == v1end)
	    break;
    }
    else if (v1val > v2val) {
	*w++ = v2val;
	v2++;
	if (v2 == v2end)
	    break;
	v2val = *v2 & mask;
    }

Update: I've tried to rewrite the routine in assembly with cmovs.
The main loop is as follows:

	mov (V2),V2val
	mov (V1),V1val
	and Mask,V2val
    L(merge):
	cmp V2val,V1val
	cmovb V1val,V2val
	lea 4(V2),Addr
	cmovae Addr,V2
	lea 4(V1),Addr
	mov V2val,(W)
	cmovbe Addr,V1
	add $4,W
	cmp V2,V2end
	je L(break)
	mov (V2),V2val
	mov (V1),V1val
	and Mask,V2val
	cmp V1,V1end
	jne L(merge)

On Haswell, this results in but a minor improvement: 11.9 cycles per
integer (downsample.c) vs 11.1 cpi (asm).  The result is much more
noticeable on Ryzen 2400G, roughly 12 cpi down to 8.  It appears that
older uarches (such as Haswell) have relatively slow cmovs (2 uops,
2c latency).  Newer uarches (Zen and Skylake) have faster cmovs (1 uop,
1c latency) with improved throughput (can execute 2 cmovs in parallel).