rpmsetcmp.c: initial revision
In rpmss/rpmsetcmp.c, I once tried to do as little work as possible by
handling each case separately.  For example, if Provides is to be
downsampled, then there is no need to install sentinels right after
decoding (as the sentinels will be installed again after downsampling).
This approach results in a combinatorial mini-explosion - the same code
gets duplicated multiple times, with only minor variations.  To cope
with the complexity, I invented a mini-language atop of cpp macros,
with the last argument in each action-like macro being a "continuation",
what to do next.  Thus I was able to combine a few actions into a
sequence in a very succinct way.

In retrospect, this seems to be of dubious value (although still
interesting).  Installing the sentinels should take only two
instructions:

	pcmpeqd %xmm0,%xmm0
	movdqu %xmm0,(Pv,Pn,4)

Further only about 2% rpmsetcmp calls need downsampling.  So saving 2
instructions 2% of the time, at the expense of duplicating the code,
does not seem to be the right goal.  Further note that most Provides
versions come from the cache, where the sentinels have to be installed
anyway.

Another issue is that we should combine two allocations that cannot be
carried out on the stack into a single malloc call.  Specifically when
we need to apply more than one downsampling step to a Provides from the
cache, we need two temporary arrays.  The second malloc call should be
avoided even if it happens 1% of the time or less.

So now I take a more practical approach:
1) There is a "fast path" for the case without downsampling, where
I still try hard to do no extra work.
2) The case with downsampling is factored into a separate function.
Some information is lost because of this - e.g. it takes "const *Pv",
so it can't clobber Pv even if Pv was decoded on the stack rather than
obtained from the cache.  But because only short versions are decoded on
the stack, requesting extra stack space should be cheap.  In other
words, this case does not interfere with malloc.
3) The case for double-sized array is handled explicitly.  This is only
applicable to Provides, as Requires can always be clobbered.

Each rpmsetcmp call on average takes about 6900 cycles (inclusive).
I tried another variant in which Requires was decoded first - it took
7000+ cycles, with perf-stat showing significantly more LLC-load-misses.
(This may be mostly due to memory layout in my test program - getline(3)
allocates P and R in a single chunks, with R placed right after P.)