downsample by right shift
It dawned on me that downsampling with a right shift instead of a mask
has two crucial advantages:
1) The order is preserved, no need to merge, so the loop becomes
trivial (and runs 4x faster in practice).  It is somewhat similar to
reducing precision of a floating-point number (by keeping only a few
most significant bits of its mantissa).
2) Any number of downsampling steps can be achieved in a single call.
This obviates the need for dealing with double-sized buffers.
Furthermore this makes inplace downsampling possible.

The downside is that downsampling won't work with old set-versions when
they are simply converted to the new format (which is currently my main
testing and benchmarking method).  They have to be regenerated from
scratch, by taking the more significant hash bits.