downsample.c: instantiate for k=1
This is a low-hanging fruit: it reduces the average cost 2.8 -> 2.4
cycles per integer with gcc9, 2.4 -> 1.9 with clang7 (on Haswell).

I also tried to unroll the loop, but clang is already good at this,
so I couldn't beat clang.  gcc is not as good, and unrolling the loop
in a clever way helps a bit, 2.4 -> 2.0.  But I decided I don't want to
interfere with the compiler in this way.  Critical loops can be written
in assembly, but this is not a critical one.