setcmploop-x86_64.S: initial revision
This is a version from a month ago, as is.  (It need to be "polished up
a bit", but each polishing step has to be carefully considered.)  I still
have doubts whether hand-written assembly on this scale is even justified,
or should rather be avoided.  (I have no clue how to address the Intel JCC
erratum, some jump instructions should not cross a 32-byte boundary.)

This version is based on gcc output.  Specifically the construction
with L(lt) in the middle of the loop is counter-intuitive, L(lt) at
the beginning is more straightforward and should work just as well.
I double-checked that L(lt) in the middle works slightly better.

This version beats gcc:
-setcmploop: 2453 cycles per call
+setcmploop: 2398 cycles per call

I'm not adding any .align directives yet.