setcmploop-x86_64.S: two loops, select step=2 vs step=4
When (Pn / 32 >= Rn), there is now the second loop with step=4.
On Haswell, this results in a sizable improvement:
-setcmploop: 2398 cycles per call
+setcmploop: 2310 cycles per call

A similar improvement on Ryzen 2400G.  In fact, always selecting step=4
is almost as good on Ryzen.  This is because cmov is cheap on Ryzen, and
bisecting with 2 cmovs is fast.  On Haswell, cmovs are not as cheap, so
a few more loop steps followed by a single cmov work better.