setcmploop-x86_64.S: optimize L(break) and return
Normally we break out of the loop in L(eq), when the last Rval from Rv
has been found in Pv, and the next Rv is at Rend.  callgrind --auto=yes
confirms this:

        .  L(breakP):
      180       xor %eax,%eax
      180       cmp Rv,Rend
      180       cmovne %eax,Ge
      180       jmp L(ret)
        .  L(breakR):
   83,719       xor %eax,%eax
   83,719       cmp Pv,Pend
   83,719       cmovne %eax,Le

This means the only case that should be optimized is breaking out of
L(eq).  Other cases can be optimized more aggerssively, but we shouldn't
do this because they are rare.  (E.g. it seems that we can immediately
return -2 for some L(breakP) but not for others.)

Then there are 4 possible return values: 0, 1, -1, -2.  They are packed
into an integer ($0x01fffe), the integer is right-shifted and its least
significant byte gets sign-extended.  If Ge and Le are both 0, the shift
amount should be 0 (so that -2 is returned).  Otherwise the integer
should be shifted by 8, 16, or 24.  Thus the shift amount can be
calculated with a single lea instruction.

I've run a more elaborate benchmark that compares setcmploop.c vs
setcmploop-x86_64.S.  (There are 99 runs, each run executes 33 full
unmet passes, 83899 rpmsetcmp calls per pass.  Each run outputs a single
number, average cycles per setcmploop call.  The statistics is over
those 99 numbers.)

On Haswell:
-setcmploop: 2452±5 cycles per call, quartiles: 2449, 2451, 2456
+setcmploop: 2304±6 cycles per call, quartiles: 2299, 2304, 2308

On Ryzen 2400G:
-setcmploop: 2159±33 cycles per call, quartiles: 2140, 2155, 2165
+setcmploop: 2014±37 cycles per call, quartiles: 1990, 2002, 2040

perf report:
    33.29%  a.out    a.out               [.] setcmploop
    29.18%  a.out    a.out               [.] decode10
    12.69%  a.out    libc-2.30.so        [.] __memcmp_sse4_1