setcmploop.md: initial revision, naive loop
It occurred to me that it may be not at all obvious why the code in
setcmploop.c works, or what it even does.  Someone like Andrew Savchenko
may complain that the whole thing is simply undocumented.

Instead of riddling the code with comments, I decided to try and write
a few paragraphs in plain English.  Explaining the idea is more
important that commenting on code details.  So it's also an experiment
in how code should be documented.  I'm not sure I like the result,
it's too didactic (there are trivial definitions and statements,
sometimes I restate the same thing "in other words", etc.)  After giving
it some thought, I believe it's not too bad.  The problem I'm actually
dealing with is human reasoning and understanding.  Those are
non-technical phenomena.  Human understanding can be bolstered up with
restating the same things "in other words", working out trivial
examples, etc.  So going a bit didactic instead of completely technical
is perhaps okay.