base64unpack-simd.h: introducing VSHLV16 for unpack10x9c15
In unpack10x9c15, bits are first spread across two V32x4 registers,
then processed and stored.  We can do better if we process both parts
in a single register, then spread and store.  This is possible if all
m-bit segments do not span longer than 2 bytes, that is, if each m-bit
segment can be placed into a 16-bit lane.  This is trivially possible
with m=8 and m=9, because there is no way 9 bits can overflow into the
third byte.  This is also possible with m=10, because it has a common
factor with 8 (if we consider 2-bit units, there is no way a 5-unit
segment can overflow into the third 4-unit segment).  This isn't
possible with m=11.

Though this can seemingly save only one instruction, the speed-up is
noticeable: we emulate variable-length left shift with multiplication,
and 32-bit vs 16-bit mul instructions have different complexities and
latencies: pmulld 10c vs pmullw 5c.  So we've traded two expensive
instructions for one much cheaper instruction.

Now, AVX2 has a cheaper variable-length 32-bit shift (vpsllvd) but there
is no 16-bit counterpart (vpsllvw), which is available only with AVX512.
So the result is not necessarily better on AVX2.  But it isn't clearly
worse, as vpsllvd is 3c on Ryzen.  (Another reason why 16-bit lanes
should be preferred is that 1 can be added to all deltas with just one
instruction.  This is better than adding 1 repeatedly in the q-delta
loop.)

-unpack10x9: 1.70 cycles per integer
+unpack10x9: 1.37 cycles per integer

bench-unmet-sse4:
-9.31 cycles per integer
+9.20 cycles per integer