setstring-dec.c: unroll q-delta loop, latency reduced 6->5
This incorporates some of my findings from [qdeltaloop.gist].
The latency of the loop-carried data dependency can be reduced 6->5,
assuming that ctz32(b) and b >>= 1 can execute in parallel.

	z = ctz32(b), z++, b >>= z;		    (3+1+2)
	z = ctz32(b), b >>= 1, b >>= z;		    (3+0+2)

I have another version of this loop which uses table lookup, it performs
better on synthetic tests (in a tight loop), but it does not seem to be
any faster in tandem with SIMD unpacking.

[qdeltaloop.gist] https://gist.github.com/svpv/b5d57dcf52e991127fca9fc06bf21b1b

Haswell:
-9.19 cycles per integer
+8.83 cycles per integer

A57:
-1.97 cycles per integer
+1.52 cycles per integer

$ zstd -d <conv.setcmp.zst |awk 'length($1)>=1024{print$1,$1}' |./bench-unmet-sse4
-8.34 cycles per integer
+7.85 cycles per integer