conv.c: increase bpp with +1
I need this to profile the decoding routine for m=11.  The order in v[]
is preserved, the original value is shifted v[i] << k, with k low
pseudo-random bits generated from v[i].  This won't work properly with
downsampling, but downsmapling doesn't work with the old data anyway
(because I changed the downsampling scheme to strip LSB instead of MSB).