introducing pack11x8c16e8, replenish the bitstream with more bits
So we've got two 11x8 versions, c15e2 and c16e8 (the second one returns
more bits to the bitstream, but the bitstream has to be refilled with
kq=4 instead of kq=5).  Which one is better?  With both versions, the
bitstream can hold at most 31 bits (at any one time).  But in fact
the second one is noticeably better: because it returns more bits to the
bitstream, the bitstream can be refilled less often.

$ zstd -d <conv+1.setcmp.zst |awk 'length($1)>=1024{print$1,$1}' |./bench-unmet-sse4
-8.54 cycles per integer
+8.37 cycles per integer

perf stat branch-misses:
-2.35% of all branches
+1.94% of all branches

A57 (bench-base64unpack-scalar):
-1.69 cycles per integer
+1.48 cycles per integer