combine extra bits differently, b << ke instead of e << bfill
On x86, this saves shl %cl which is 2c (while shl $imm8 is only 1c).
This also solves another problem: shl %cl interfering with register
allocation.  If you need to shift by bfill, it's best to place bfill
into the %ecx register.  But in the q-delta loop, you need to read from
the variable-length bitstream (with shr %cl discarding the bits), so
%ecx will be clobbered.

Strangely enough, this seems to help even on ARM, where shifts by
immediate or by any register should be equally cheap.  Which reminds me
of a horseshoe, which brings you luck whether you believe in it or not.

$ zstd -d <conv+1.setcmp.zst |awk 'length($1)>=1024{print$1,$1}' |./bench-unmet-sse4
-8.37 cycles per integer
+8.24 cycles per integer

A57 (bench-base64unpack-scalar):
-1.48 cycles per integer
+1.45 cycles per integer