VMOVZW{LO,HI} for unpack12x8c16, do not interleave
Interleaving in pack12x8c16 hampers LZ compression.  The interleaved
format is 04152637.  It is not immediately obvious why, if we have two
similar set-strings (with only a few different elements), we are more
likely to find a match between the sequence of adjacent 2-byte deltas
(0123...) rather than interleaved deltas (0415...).  After all, the
hash output for deltas is essentially random.  Rearranging them in
different ways can't make them look more random or less random.

The explanation is that interleaving makes LZ matches sensitive to the
offset.  Say we want to match

    01234567 8...
    a0123456 7...

With interleaving, we get

    04152637 8...
    a3041526 7...

So we've got a shorter match (6 deltas instead of 7 within the first
block), and the match gets interrupted at the end of each block.  Note
that we are likely to have two adjacent 12x8 blocks, because we refill
the bitstream with 30 bits which cover 30/8/2=1.875 blocks on average.

With this commit, the compressed size of the working set converted
with +2 (m=12 for most Provides versions) is very close to that of m=11.
Unlike m=11, packed m=12 deltas are on a byte boundary, LZ is good at it.

$ zstd -d <conv+2.setcmp.zst |awk 'length($1)>=1024{print$1,$1}' |./bench-unmet-sse4
-7.71 cycles per integer
+7.61 cycles per integer

$ ls -1s conv+*zst*
41068 conv+1.setcmp.zst	    (m=11)
41744 conv+2.setcmp.zst	    (m=12 adjacent, this commit)
44252 conv+2.setcmp.zst-    (m=12 interleaved, previous version)